#26283 Move Packstation logic to Trait
Organize functionality in specific Trait
