+++ src/Objects/PickupPoint.php
@@ -4,6 +4,7 @@ namespace Monta\CheckoutApiWrapper\Objects;
 
 // alias for sibling must remain or not all autoloading will work
 use Monta\CheckoutApiWrapper\Objects\Option as Option;
+use Monta\CheckoutApiWrapper\Traits\Packstations;
 
 /**
  * Class PickupPoint
@@ -11,6 +12,8 @@ use Monta\CheckoutApiWrapper\Objects\Option as Option;
  */
 class PickupPoint extends Option
 {
+    use Packstations;
+
     public const string PICKUP_OPTIONS_KEY = 'PickupOptions';
 
     public const string PICKUP_STORE_KEY = 'StoreLocation';
@@ -374,33 +377,4 @@ class PickupPoint extends Option
             . ' | ' . $this->getDistanceMeters() . 'km';
     }
 
-    /**
-     * Field contains o.a. the Postnumber for DHLDE packstations
-     *
-     * @return string
-     */
-    protected function getShipperOptionsWithValue(): string
-    {
-        return $this->shipperOptionsWithValue;
-    }
-
-    /**
-     * @param string $addValue - New shipperoption to add to CSV string
-     * @return void
-     */
-    public function addShipperOptionsWithValue(string $addValue): void
-    {
-        // Get property from CSV string as array
-        $codes = array_filter(
-            array_map('trim', explode(',', $this->shipperOptionsWithValue ?? '')),
-        );
-
-        // TODO every change now adds a new code to line. Refactor to always just write "Postnumber_XX,{postnumber}"
-        if (!in_array($addValue, $codes, true)) {
-            $codes[] = $addValue;
-        }
-
-        // Implode back into CSV string
-        $this->shipperOptionsWithValue = implode(',', $codes);
-    }
 }
